Fixed compilation warnings
git-svn-id: http://miranda.googlecode.com/svn/trunk@12190 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/Gadu-Gadu/libgadu/libgadu.c
@@ -2262,7 +2262,7 @@ int gg_typing_notification(struct gg_session *sess, uin_t recipient, int length)
 	struct gg_typing_notification pkt;
 	uin_t uin;
 
-	pkt.length = gg_fix16(length);
+	pkt.length = gg_fix16((uint16_t)length);
 	uin = gg_fix32(recipient);
 	memcpy(&pkt.uin, &uin, sizeof(uin_t));
